pci: issue a full SMP memory barrier before setting dbs
In main.c's nvmev_proc_dbs(), several admin functions are called if dbs
differs from an older state.

If dbs is committed to memory before others (e.g., nvme_sq[], nvme_cq[]),
it can cause a malfunction.

A lock between nvmev_proc_dbs() and nvmev_proc_bars() would solve this
entirely, but it's probably an overkill. Instead, issue a full SMP memory
barrier so that the updates are seen in order with regards to dbs.

Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>