admin: implement NVME_LOG_CMD_EFFECTS for ZNS
The Linux's NVMe device driver queries NVME_LOG_CMD_EFFECTS [1] when ids.csi
is set, which is the case for ZNS [2].

If a vague buffer is returned (in this case, a page filled with 0xFFFFFFFF),
it results in system malfunction.

Fix this by correctly implementing NVME_LOG_CMD_EFFECTS per the current
NVMeVirt's ZNS capabilities.

Link: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/drivers/nvme/host/core.c?h=v6.1#n4029 [1]
Link: https://github.com/snu-csl/nvmevirt/blob/7d3e1c821e7512ece063555cf8720a08a89997ff/zns_ftl.c#L126 [2]
Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>