admin: make it abundantly clear that unimplemented lid will result in errors
The NVMe protocol mandates several commands (lid) to be implemented, but some
aren't in NVMeVirt.

As the NVMe host device driver will always assume that the device will return
the correct values, blindly memset'ing the return buffer will always result in
heavy system malfunction due to incorrect memory dereferences.

Warn the users and make it perfectly clear that this needs to be implemented.

Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>