zns: fix BLKREPORTZONE ioctl on start_zid > 0
Userspace may ask for BLKREPORTZONE with start_zid > 0 when a virtually
mapped device accesses the middle of the actual device.

When that happens, NVMeVirt triggers a buffer overflow and wrongfully
copies to `report->zd` from `zone_descs[]` with a length that exceeds its
size. The Linux kernel subsequently errors out on the ioctl() call as
invalid ZNS type is detected.

To fix this, adjusting the user's nr_zones directly is required.

Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>